security merge of master into readme

+++ backend/controllers/board.controller.js
@@ -2,23 +2,19 @@ const Board = require("../models/board.model.js");
 const Card = require("../models/card.model.js");
 const User = require("../models/user.model.js");
 
-const getBoards = async (req, res) => {
-  try {
-    const boards = await Board.find({});
-    res.status(200).json(boards);
-  } catch (err) {
-    res.status(500).json({ message: "Erro ao buscar boards", error: err });
-  }
-};
-
-//TO DO REGEX
 const getBoardById = async (req, res) => {
+  const userId = req.userId
   try {
     const board = await Board.findById(req.params.id);
     if (!board) {
       return res.status(404).json({ message: "Board não encontrado" });
     }
-    res.status(200).res.send(board);
+
+    if (board.createdBy.toString() !== userId && !board.sharedWith.includes(userId)) {
+      return res.status(401).json({ message: "Acesso negado" });
+    }
+
+    res.status(200).json(board);
   } catch (err) {
     res.status(500).json({ message: "Erro ao buscar board", error: err });
   }
@@ -50,13 +46,21 @@ const createBoard = async (req, res) => {
 
 const updateBoard = async (req, res) => {
   try {
-      const id = req.params.id;
-      const board = await Board.findByIdAndUpdate(id, req.body, { new: true });
-      if (!board) {
-          res.status(404).json({ message: 'Board não encontrado' });
-          return;
-      }
-      res.status(200).json(board);
+    const userId = req.userId
+    const board = await Board.findById(req.params.id);
+    if (!board) {
+      return res.status(404).json({ message: "Board não encontrado" });
+    }
+
+    if (board.createdBy.toString() !== userId && !board.sharedWith.includes(userId)) {
+      return res.status(401).json({ message: "Acesso negado" });
+    }
+
+    const id = req.params.id;
+    const updateBoard = await Board.findByIdAndUpdate(id, req.body, { new: true });
+    const updatedBoard = await Board.findById(req.params.id);
+
+    res.status(200).json(updateBoard);
 
   } catch (error) {
       res.status(500).json({ error: error.message });
@@ -65,15 +69,23 @@ const updateBoard = async (req, res) => {
 
 const deleteBoard = async (req, res) => {
   try {
+    const userId = req.userId
+    const board = await Board.findById(req.params.id);
+    if (!board) {
+      return res.status(404).json({ message: "Board não encontrado" });
+    }
+
+    if (board.createdBy.toString() !== userId && !board.sharedWith.includes(userId)) {
+      return res.status(401).json({ message: "Acesso negado" });
+    }
+    
     await Card.deleteMany({ board: req.params.id });
     await User.updateMany(
       { board: req.params.id },
       { $pull: { board: req.params.id } }
     );
-    const board = await Board.findByIdAndDelete(req.params.id);
-    if (!board) {
-      return res.status(404).json({ message: "Board não encontrado" });
-    }
+    const deleted_board = await Board.findByIdAndDelete(req.params.id);
+
     res.status(200).json({ message: "Board excluído com sucesso" });
   } catch (err) {
     res.status(500).json({ message: "Erro ao excluir board", error: err });
@@ -83,7 +95,7 @@ const deleteBoard = async (req, res) => {
 const getBoardsByUserId = async (req, res) => {
   try {
     const userId = req.userId
-    
+
     const boards = await Board.find({
       $or: [
         { createdBy: userId }, 
@@ -97,12 +109,96 @@ const getBoardsByUserId = async (req, res) => {
   }
 }
 
+const addContribuitorToBoard = async (req, res) => {  
+  try {
+    const { addedUserEmail } = req.body;
+    const { boardId } = req.params
+    const userId = req.userId
+
+    const board = await Board.findById(boardId);
+    if (!board) {
+      return res.status(404).json({ message: "Board não encontrado" });
+    }
+
+    if (board.createdBy.toString() !== userId && !board.sharedWith.includes(userId)) {
+      return res.status(401).json({ message: "Acesso negado" });
+    }
+
+    const addedUser = await User.findOne({ email: addedUserEmail })
+    if (!addedUser) {
+      return res.status(404).json({ message: "Usuário não encontrado" });
+    }
+    
+    const addedUserId = addedUser._id
+
+    if (board.createdBy.toString() == addedUserId.toString()) {
+      return res.status(400).json({ message: "Usuário já é proprietário do board" });
+    }
+
+    if (board.sharedWith.includes(addedUserId)) {
+      return res.status(400).json({ message: "Usuário já é contribuidor do board" });
+    }
+
+    // Atualizar o board para adicionar o userId no shareWith
+    const updatedBoard = await Board.findByIdAndUpdate(
+      boardId,
+      { $addToSet: { sharedWith: addedUserId } }, // $addToSet evita duplicados no array
+      { new: true } // Retorna o board atualizado
+    );
+    res.status(200).json({ message: "Usuário adicionado ao board com sucesso" });
+  } catch (err) {
+    res.status(500).json({ message: "Erro ao adicionar usuário ao board", error: err });
+  }
+}
+
+const removeContribuitorFromBoard = async (req, res) => {  
+  try {
+    const { removedUserEmail } = req.body;
+    const { boardId } = req.params
+    const userId = req.userId
+
+    const board = await Board.findById(boardId);
+    if (!board) {
+      return res.status(404).json({ message: "Board não encontrado" });
+    }
+
+    if (board.createdBy.toString() !== userId && !board.sharedWith.includes(userId)) {
+      return res.status(401).json({ message: "Acesso negado" });
+    }
+
+    const removedUser = await User.findOne({ email: removedUserEmail })
+    if (!removedUser) {
+      return res.status(404).json({ message: "Usuário já não tem acesso ao board" });
+    }
+    
+    const removedUserId = removedUser._id
+
+    if (board.createdBy == removedUserId) {
+      return res.status(400).json({ message: "O proprietário do board não pode ser removido" });
+    }
+
+    if (!board.sharedWith.includes(removedUserId)) {
+      return res.status(400).json({ message: "Usuário já não é contribuidor do board" });
+    }
+
+    // Atualizar o board para adicionar o userId no shareWith
+    const updatedBoard = await Board.findByIdAndUpdate(
+      boardId,
+      { $pull: { sharedWith: removedUserId } }, // 
+      { new: true } // Retorna o board atualizado
+    );
+    res.status(200).json({ message: "Usuário removido do board com sucesso" });
+  } catch (err) {
+    res.status(500).json({ message: "Erro ao adicionar usuário ao board", error: err });
+  }
+}
 
 module.exports = {
-  getBoards,
   getBoardById,
   createBoard,
   updateBoard,
   deleteBoard,
   getBoardsByUserId,
+  addContribuitorToBoard,
+  removeContribuitorFromBoard,
 };
