+++ backend/controllers/card.controller.js
@@ -3,7 +3,8 @@ const Board = require("../models/board.model.js");
 
 const createCard = async (req, res) => {
   try {
-    const { title, description, status } = req.body;
+    const { title, description } = req.body;
+    var { status } = req.body;
     const { boardId } = req.params;
     const userId = req.userId;
     
@@ -38,21 +39,17 @@ const createCard = async (req, res) => {
   }
 };
 
-const getCards = async (req, res) => {
-  try {
-    const cards = await Card.find({});
-    res.status(200).json(cards);
-  } catch (error) {
-    res.status(500).json({ error: error.message });
-  }
-};
-
 const getCardsByBoard = async (req, res) => {
   try {
+    const userId = req.userId;
     const { boardId } = req.params;
-    const board = await Board.findById(boardId);
+    const board = await Board.findById(boardId)
     if (!board) {
       return res.status(404).json({ message: "Board não encontrado" });
+    }
+
+    if (board.createdBy.toString() !== userId && !board.sharedWith.includes(userId)) {
+      return res.status(403).json({ message: "Você não tem permissão para visualizar os cards deste board!" });
     }
 
     const cards = await Card.find({ board: boardId });
@@ -64,7 +61,19 @@ const getCardsByBoard = async (req, res) => {
 
 const updateCard = async (req, res) => {
   try {
+    const userId = req.userId;
     const { id } = req.params;
+
+    const card = await Card.findById(id);
+    if (!card) {
+      return res.status(404).json({ message: "Card não encontrado" });
+    }
+
+    const board = await Board.findById(card.board);
+
+    if (board.createdBy.toString() !== userId && !board.sharedWith.includes(userId)) {
+      return res.status(403).json({ message: "Você não tem permissão para editar este card!" });
+    }
 
     const updated = await Card.findByIdAndUpdate(id, req.body, { new: true });
 
@@ -80,18 +89,19 @@ const updateCard = async (req, res) => {
 
 const deleteCard = async (req, res) => {
   try {
+    const userId = req.userId;
     const { id } = req.params;
     const cardExists = await Card.findById(id);
     if (!cardExists) {
       return res.status(404).json({ message: "Card não encontrado" });
     }
 
-    const boardsUpdated = await Board.updateMany(
-      { cards: id },
-      { $pull: { cards: id } } // Remove o ID do Card do array
-    );
+    const board = await Board.findById(cardExists.board);
 
-    console.log(`Boards atualizados: ${boardsUpdated.modifiedCount}`);
+    if (board.createdBy.toString() !== userId && !board.sharedWith.includes(userId)) {
+      return res.status(403).json({ message: "Você não tem permissão para deletar este card!" });
+    }
+
     await Card.findByIdAndDelete(id);
 
     res.status(200).json({
@@ -104,7 +114,6 @@ const deleteCard = async (req, res) => {
 
 module.exports = {
   createCard,
-  getCards,
   updateCard,
   deleteCard,
   getCardsByBoard
